fix: Move storage/event/sender to keboolasink
Add test for event sender. Adapt it with error pointer. Remove unused
tests with gcs and s3 writers.
